+++ src/interfaces/IRewardsHandler.sol
@@ -3,23 +3,22 @@
 pragma solidity >=0.8.0;
 
 interface IRewardsHandler {
-
     struct UserBalance {
-        uint allocPoint; // Allocation points
-        uint lastMul;
+        uint256 allocPoint; // Allocation points
+        uint256 lastMul;
     }
 
-    function receiveFee(address token, uint amount) external;
+    function receiveFee(address token, uint256 amount) external;
 
-    function updateLPShares(uint fnftId, uint newShares) external;
+    function updateLPShares(uint256 fnftId, uint256 newShares) external;
 
-    function updateBasicShares(uint fnftId, uint newShares) external;
+    function updateBasicShares(uint256 fnftId, uint256 newShares) external;
 
-    function getAllocPoint(uint fnftId, address token, bool isBasic) external view returns (uint);
+    function getAllocPoint(uint256 fnftId, address token, bool isBasic) external view returns (uint256);
 
-    function claimRewards(uint fnftId, address caller) external returns (uint);
+    function claimRewards(uint256 fnftId, address caller) external returns (uint256);
 
     function setStakingContract(address stake) external;
 
-    function getRewards(uint fnftId, address token) external view returns (uint);
+    function getRewards(uint256 fnftId, address token) external view returns (uint256);
 }
