+++ src/lib/IWETH.sol
@@ -3,10 +3,8 @@
 pragma solidity ^0.8.0;
 
 interface IWETH {
-
     function deposit() external payable;
     // Introduced later in development
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
-
+    function transfer(address to, uint256 value) external returns (bool);
+    function withdraw(uint256) external;
 }
